changed: put videodatabase create and clean methods in separate file
